All code completion events are now routed via ServiceProviderManager this allows CodeLite better control of the order of execution by the priority provided by the ServiceProvider classes. Don't use thread_local :( this is due to MinGW compiler bug on Windows. See this report for example: https://github.com/msys2/MINGW-packages/issues/2519 Make FileExtManager thread safe by adding a critical section

+++ LiteEditor/threadlistpanel.cpp
@@ -65,7 +65,7 @@ void ThreadListPanel::PopulateList(const ThreadEntryArray& threads)
         // No need to repopulate the list, just set the active thread indicator
 
         // Loop over the table and set all threads to "NO"
-        for(int i = 0; i < m_dvListCtrl->GetItemCount(); ++i) {
+        for(size_t i = 0; i < m_dvListCtrl->GetItemCount(); ++i) {
             ThreadListClientData* d = (ThreadListClientData*)m_dvListCtrl->GetItemData(m_dvListCtrl->RowToItem(i));
             d->GetThreadEntry().active = false;
             wxVariant v = "NO";
@@ -86,7 +86,7 @@ void ThreadListPanel::PopulateList(const ThreadEntryArray& threads)
 
         if(threadID != wxNOT_FOUND) {
             // Update the new active thread
-            for(int i = 0; i < m_dvListCtrl->GetItemCount(); ++i) {
+            for(size_t i = 0; i < m_dvListCtrl->GetItemCount(); ++i) {
                 ThreadListClientData* d = (ThreadListClientData*)m_dvListCtrl->GetItemData(m_dvListCtrl->RowToItem(i));
                 if(d->GetThreadEntry().dbgid == threadID) {
                     d->GetThreadEntry().active = true;
@@ -136,7 +136,7 @@ void ThreadListPanel::PopulateList(const ThreadEntryArray& threads)
 
 void ThreadListPanel::Clear()
 {
-    for(int i = 0; i < m_dvListCtrl->GetItemCount(); ++i) {
+    for(size_t i = 0; i < m_dvListCtrl->GetItemCount(); ++i) {
         ThreadListClientData* d = (ThreadListClientData*)m_dvListCtrl->GetItemData(m_dvListCtrl->RowToItem(i));
         delete d;
     }
